Add timeline, template, and initial features
To make sure we stick to something, the proposal here is that we start
by just trying to pile up a bunch of data, and then assess what data
format we should use and where it should live.

It would be great if we could be ready to discuss the data format by
TPAC so that we can also discuss this with other members of the W3C.

We can then try commit to our one year goal and lock in the data format
by Feb 2025.