Move platform descriptions to behavior field
This helps with the styling because we can then display the
descriptions inline with the various WebView providers.

This also means that the behaviour data is stored within
our data structure. It does mean that we are not fully compliant
with caniuse's data structure but as long as we maintain a
superset, we should be fine on consuming that data for our
own data sources.