Add bcd features to features search results
We can query the a cdn to retrieve a big ol' json full of
the bcd.

From that we are able to generate search results.
Using a custom plugin to do this.

We will need to change the gitbug action to deploy
from a github action but it looks like this is fairly easy.
The reason we need to change this is because the github pages gem
doesn't allow us to use custom plugins.

Once this is up, we can iterate over it and integrate the data into
the rest of the site a little nicer.