Add plugin to generate WebView baseline results
It would be super handy to get a better sense of
what WebView features will both impact baseline,
and just generally where we are at.

The web features repo is very node friendly
so it seems easier to work in that.
For that reason, I'm being a little funky
and running a jekyll plugin that just executes a
node script to create an untracked data file
with our baseline results.

The important bit of this change is getting the
data in a structured way we can work with
so I've added a page that isn't linked from anywhere
else with the results but I haven't styled it for the moment.